Benchmarking applications
Written in C, to benchmark the CPU performance as well as the Memory, Disk read write speeds of a machine hardware with and without multi-threading alongside AVX instruction set usage to further enhance the performance. 
In addition, also performed the network speed benchmarking on the loopback address as well as between 2 different nodes. 
GPU benchmarking using CUDA for Nvidea video cards. All of the mentioned run on Chameleon cloud server nodes.